3点リーダーを押すとメニューが出てくるよう改修(出したメニューは再度3点リーダを押すと消える。今後、メニューの外側を押すと消えるよう改修予定)
